[ref] Extract notifications into separate components
